# RowwiseMoments: use float as acc type for bfloat16 inputs