@mingfeima 2022-07-21
This PR is to fix https://github.com/pytorch/pytorch/issues/77507

Allowing bfloat16 to be accumulated in float32 also brings performance improvement since we don't have to redundant dtype conversion which is very time consuming.

* **opt1** refers to https://github.com/pytorch/pytorch/pull/81849
* **opt2** refers to https://github.com/pytorch/pytorch/pull/81850
* **diff1** refers is improvement of opt1 against before
* **diff2** refers is improvement of opt2 against opt1

this PR has no effect on fp32 performance, will bring `4-5x` performance improvement on bf16.

### avx512 result

number of cores | data type | before | opt1 | opt2 | diff1 | diff2
-- | -- | -- | -- | -- | -- | --
20 | float32 | 0.439 | 0.210 | 0.215 | 209.05% | 97.67%
20 | bfloat16 | 2.479 | 0.770 | 0.178 | 321.95% | 432.58%
1 | float32 | 6.308 | 2.661 | 2.618 | 237.05% | 101.64%
1 | bfloat16 | 39.765 | 12.267 | 2.309 | 324.16% | 531.27%

### avx2 result

number of cores | data type | before | opt1 | opt2 | diff1 | diff2
-- | -- | -- | -- | -- | -- | --
12 | float32 | 1.248 | 0.540 | 0.527 | 231.11% | 102.47%
12 | bfloat16 | 8.487 | 2.030 | 0.458 | 418.08% | 443.23%
1 | float32 | 10.792 | 4.349 | 4.416 | 248.15% | 98.48%
1 | bfloat16 | 66.366 | 19.252 | 3.524 | 344.72% | 546.31%