@malfet 2022-08-10
One can rebase using the rebase command of the mergebot