@malfet 2022-08-11
@pytorchbot revert -c weird -m "Revert as caused perf regression, see https://github.com/pytorch/benchmark/issues/1099"